refactor and categorize utils folder
