+++ utils/index.js
@@ -1,0 +1,13 @@
+import { addExerciseDataToLoggedWorkout } from "./dataMutators";
+import { addExerciseDataToWorkout } from "./dataMutators";
+import { getCurrYearMonthDay } from "./dateAndTime";
+import { timeSince } from "./dateAndTime";
+import { round } from "./math";
+
+export {
+  addExerciseDataToLoggedWorkout,
+  addExerciseDataToWorkout,
+  getCurrYearMonthDay,
+  timeSince,
+  round,
+};
