+++ utils/math.js
@@ -1,0 +1,4 @@
+export const round = (value, precision) => {
+  var multiplier = Math.pow(10, precision || 0);
+  return Math.round(value * multiplier) / multiplier;
+};
